src/github/requirements.txt
# runtime
PyGithub==2.3.0
PyYAML==6.0.2
boto3==1.34.162
requests==2.32.3
pynacl==1.5.0

# tighten transitive that triggers some scanners
urllib3>=2.2.2

# dev-only type stubs (safe to keep here; harmless in prod)
boto3-stubs[ssm]==1.34.162
types-PyYAML==6.0.12.20240808
